refactor: split dispatch.ts into domain files with shared lock helper
1510-line monolith split into 6 domain files:
- dispatch-lock.ts: withParallelLock helper + getMaxParallelAgents
- dispatch-triage.ts: dispatchTriageBatch
- dispatch-dev.ts: investigate, spec-author, fix-issue, parallel-implement, resolve-conflict
- dispatch-qa-review.ts: qa, review, retro, needs-fix, qa-failed
- dispatch-discover.ts: grill-and-prd, retry-write-prd, revise-prd, decompose-prd
- dispatch-routing.ts: dispatchForLabel, dispatchForIssue, dispatchResumeIssue

dispatch.ts is now a barrel re-export — no callers change.

withParallelLock collapses 13x repeated isInFlight/tryAcquire/release/drain
boilerplate. Post-lock callback return value supports chaining to other
dispatchers that need the same lock after release (investigate→investigation-
complete, resolve-conflict→retro, qa-failed→needs-fix).

dispatchRevisePrd intentionally bypasses withParallelLock — it is user-triggered
and does not enqueue or drain on cap-full.
