+++ apps/server/src/shared/dispatch.ts
@@ -1,1509 +1,21 @@
-import { mkdirSync } from 'node:fs';
-import { tmpdir } from 'node:os';
-import { join } from 'node:path';
-import { ClaudeCliRuntime } from '@goose-hub/core/agent-runtime/claude-cli.js';
-import { resolveGlobalSettingsForProject } from '@goose-hub/core/agent-runtime/resolve-for-project.js';
-import { getUseMultiAgentPipeline } from '@goose-hub/core/db/repositories/project-settings.js';
-import { getEngineeringSpec } from '@goose-hub/core/engineering-specs/repository.js';
-import { emitStateTransitionEvent } from '@goose-hub/core/event-stream/state-transition.js';
-import { eventStore } from '@goose-hub/core/event-stream/store.js';
-import { logger } from '@goose-hub/core/logger.js';
-import { filterEligibleByDependencies } from '@goose-hub/core/projects/dependency-scheduler.js';
-import { parallelLock } from '@goose-hub/core/projects/parallel-lock.js';
-import type { StateName } from '@goose-hub/core/state-machine/states.js';
-import { createProjectAwareTargetSource } from '@goose-hub/core/state-source/dependency-resolver.js';
-import type { InvestigateOutput } from '@goose-hub/skills/investigate/schema.js';
-import { EngineeringSpecSchema } from '@goose-hub/skills/spec-author/schema.js';
-import { parseAcceptanceCriteria } from '../domains/issues/parse-acceptance.js';
-import { runRetroForItem } from '../domains/workflows/retro-batch.js';
-import { maybeFireSprintReview } from '../domains/workflows/sprint-review-trigger.js';
-import { runTriageBatch } from '../domains/workflows/triage-batch.js';
-import { getProject } from './projects.js';
-import { REPO_ROOT, sliceUrl } from './slice-url.js';
-import { getSourceForSlug } from './source.js';
-
-const _triageBatchInFlight = new Set<string>();
-const _triageBatchPending = new Set<string>();
-
-type PendingWorkflowEntry = {
-  issueNumber: number;
-  dispatchFn: (slug: string, issueNumber: number) => Promise<void>;
-};
-const _workflowPending = new Map<string, PendingWorkflowEntry[]>();
-
-function enqueueWorkflow(
-  slug: string,
-  issueNumber: number,
-  dispatchFn: (slug: string, issueNumber: number) => Promise<void>,
-): void {
-  let queue = _workflowPending.get(slug);
-  if (queue == null) {
-    queue = [];
-    _workflowPending.set(slug, queue);
-  }
-  // Dedup by (issueNumber, dispatchFn) so different workflows for the same issue can both queue.
-  if (queue.some((e) => e.issueNumber === issueNumber && e.dispatchFn === dispatchFn)) return;
-  queue.push({ issueNumber, dispatchFn });
-}
-
-function drainPending(slug: string): void {
-  const queue = _workflowPending.get(slug);
-  if (queue == null || queue.length === 0) return;
-  const entry = queue.shift();
-  if (queue.length === 0) _workflowPending.delete(slug);
-  if (entry != null) {
-    entry.dispatchFn(slug, entry.issueNumber).catch((err: unknown) => {
-      logger.error('drainPending: dispatch failed', {
-        slug,
-        issueNumber: entry.issueNumber,
-        error: String(err),
-      });
-    });
-  }
-}
-
-async function getMaxParallelAgents(slug: string): Promise<number> {
-  const cfg = await getProject(slug);
-  const settings = resolveGlobalSettingsForProject(slug, cfg?.budgets);
-  return settings.maxParallelAgents ?? 1;
-}
-
-/**
- * Centralised workflow dispatcher (#207). Domains never import from sibling
- * domains (STANDARDS.md). Webhooks, the projects router, and any other
- * caller that needs to kick off a workflow goes through these helpers.
- *
- * Slice workflows are dynamic-imported by URL because `slices/` lives at the
- * repo root (outside any workspace package) — see FACTORY_RULES rule 28a.
- * URLs are resolved via `sliceUrl()` from `./slice-url.js`.
- */
-
-/** Run the triage batch for a project. Coalesces concurrent calls per slug. */
-export function dispatchTriageBatch(slug: string): Promise<void> {
-  if (_triageBatchInFlight.has(slug)) {
-    _triageBatchPending.add(slug);
-    return Promise.resolve();
-  }
-  _triageBatchInFlight.add(slug);
-  return Promise.resolve()
-    .then(() => runTriageBatch(slug))
-    .catch((err: unknown) => {
-      logger.error('dispatchTriageBatch failed', { slug, error: String(err) });
-    })
-    .finally(() => {
-      _triageBatchInFlight.delete(slug);
-      if (_triageBatchPending.delete(slug)) {
-        void dispatchTriageBatch(slug);
-      }
-    });
-}
-
-/** Run the investigate workflow for a single issue. Drops duplicate triggers for the same issue. */
-export async function dispatchInvestigate(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchInvestigate: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchInvestigate: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchInvestigate);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runInvestigateWorkflow } = (await import(sliceUrl('investigate'))) as {
-      runInvestigateWorkflow: (
-        item: unknown,
-        source: unknown,
-        slug: string,
-        repoRoot: string,
-        deps?: Record<string, unknown>,
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchInvestigate: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    const mockInvestigateDeps: Record<string, unknown> | undefined =
-      process.env.MOCK_AGENTS === 'true'
-        ? {
-            createWorktreeImpl: () => '/mock/worktree',
-            prewarmWorktreeImpl: () => undefined,
-          }
-        : undefined;
-    await runInvestigateWorkflow(item, source, slug, REPO_ROOT, mockInvestigateDeps);
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-
-  // Chain to investigation-complete routing outside the lock. In production the
-  // GitHub label-change webhook fires dispatchForLabel('factory:investigation-complete')
-  // instead; in MOCK_SOURCE mode there are no webhooks so we chain explicitly.
-  // Idempotency is enforced by the state check inside dispatchInvestigationComplete.
-  await dispatchInvestigationComplete(slug, issueNumber);
-}
-
-/** Run the M7 fix-issue workflow for a single issue (#183). Drops duplicate triggers for the same issue. */
-export async function dispatchSpecAuthor(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchSpecAuthor: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchSpecAuthor: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchSpecAuthor);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runSpecAuthorWorkflow } = (await import(sliceUrl('spec-author'))) as {
-      runSpecAuthorWorkflow: (
-        item: unknown,
-        source: unknown,
-        slug: string,
-        repoRoot: string,
-        deps?: Record<string, unknown>,
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchSpecAuthor: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-
-    const depProjectConfig = await getProject(slug);
-    if (depProjectConfig != null) {
-      const fetchTarget = await createProjectAwareTargetSource();
-      const { eligible } = await filterEligibleByDependencies([item], {
-        currentRepo: depProjectConfig.source.repo,
-        fetchTarget,
-        source,
-      });
-      if (eligible.length === 0) {
-        logger.info('dispatchSpecAuthor: item blocked by deps, skipping', { slug, issueNumber });
-        return;
-      }
-    }
-
-    const mockDeps: Record<string, unknown> | undefined =
-      process.env.MOCK_AGENTS === 'true'
-        ? { createWorktreeImpl: () => '/mock/worktree' }
-        : undefined;
-
-    await runSpecAuthorWorkflow(item, source, slug, REPO_ROOT, mockDeps);
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-type InvCompletePayload = { investigate?: InvestigateOutput; investigationRunId?: string };
-
-/**
- * For bugs in the M19 pipeline: read the latest investigation event and decide
- * whether the fix warrants the full spec-author → parallel-build pipeline or
- * whether a single developer agent (legacy) is sufficient.
- *
- * Simple = ≤2 key files, high confidence, no open questions.
- * When no investigation exists, defaults to legacy (fail-safe).
- */
-function resolveFixIssuePipelineForBug(
-  projectId: string,
-  workItemId: string,
-): 'spec-author' | 'legacy' {
-  const [latestInv] = eventStore.replay({
-    projectId,
-    workItemId,
-    kind: 'agent.investigation-complete',
-    order: 'desc',
-    limit: 1,
-  });
-  if (latestInv == null) return 'legacy';
-  const investigate = (latestInv.payload as InvCompletePayload).investigate;
-  if (investigate == null) return 'legacy';
-  // Test files are co-changed with their source; exclude from complexity count.
-  // openQuestions excluded: agents emit informational notes there even at high confidence.
-  const nonTestFiles = investigate.keyFiles.filter((f) => !/\.(test|spec)\.[jt]sx?$/.test(f.path));
-  return nonTestFiles.length <= 3 && investigate.confidence === 'high' ? 'legacy' : 'spec-author';
-}
-
-export async function dispatchFixIssue(slug: string, issueNumber: number): Promise<void> {
-  const projectForFlag = await getProject(slug);
-  const useMultiAgent =
-    projectForFlag != null ? getUseMultiAgentPipeline(projectForFlag.id) : false;
-  logger.info('dispatchFixIssue: pipeline flag', {
-    slug,
-    issueNumber,
-    useMultiAgentPipeline: useMultiAgent,
-  });
-
-  if (useMultiAgent) {
-    const routingSource = await getSourceForSlug(slug);
-    if (routingSource == null) {
-      logger.error('dispatchFixIssue: no source for slug', { slug });
-      return;
-    }
-    const routingItem = await routingSource.getItem(issueNumber.toString());
-    if (
-      routingItem.type === 'bug' &&
-      resolveFixIssuePipelineForBug(slug, routingItem.id) === 'legacy'
-    ) {
-      logger.info('dispatchFixIssue: simple bug → legacy single-agent path', {
-        slug,
-        issueNumber,
-      });
-      // fall through to legacy path
-    } else {
-      await dispatchSpecAuthor(slug, issueNumber);
-      return;
-    }
-  }
-
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchFixIssue: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchFixIssue: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchFixIssue);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runFixIssueWorkflow } = (await import(sliceUrl('fix-issue'))) as {
-      runFixIssueWorkflow: (
-        item: unknown,
-        source: unknown,
-        slug: string,
-        repoRoot: string,
-        deps?: Record<string, unknown>,
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchFixIssue: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-
-    // Dependency gate: skip dispatch if any dep is open or unregistered.
-    // filterEligibleByDependencies applies schedule:blocked-by and needs-human
-    // labels as side-effects, so we only need to check eligible.length here.
-    const depProjectConfig = await getProject(slug);
-    if (depProjectConfig != null) {
-      const fetchTarget = await createProjectAwareTargetSource();
-      const { eligible } = await filterEligibleByDependencies([item], {
-        currentRepo: depProjectConfig.source.repo,
-        fetchTarget,
-        source,
-      });
-      if (eligible.length === 0) {
-        logger.info('dispatchFixIssue: item blocked by deps, skipping', {
-          slug,
-          issueNumber,
-        });
-        return;
-      }
-    }
-
-    const mockDeps: Record<string, unknown> | undefined =
-      process.env.MOCK_OPEN_PR === 'true'
-        ? {
-            openPRImpl: () =>
-              Promise.resolve({
-                prNumber: 999,
-                prUrl: 'https://github.com/shaunnez/goose-hub/pull/999',
-                branch: 'factory/mock-run',
-                base: 'main',
-              }),
-            createWorktreeImpl: () => '/mock/worktree',
-            prewarmWorktreeImpl: () => undefined,
-            cleanupWorktreeImpl: () => undefined,
-            resolveWorktreeHeadShaImpl: () => 'mock-sha-abc123',
-            orchestratorCommitAllImpl: () => 'mock-commit-sha',
-          }
-        : undefined;
-
-    await runFixIssueWorkflow(item, source, slug, REPO_ROOT, mockDeps);
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/** Run M19 parallel-implement for a spec-ready issue. Drops duplicate triggers. */
-export async function dispatchParallelImplement(slug: string, issueNumber: number): Promise<void> {
-  const projectForFlag = await getProject(slug);
-  const projectId = projectForFlag?.id ?? slug;
-  const useMultiAgent = projectForFlag != null ? getUseMultiAgentPipeline(projectId) : false;
-  logger.info('dispatchParallelImplement: pipeline flag', {
-    slug,
-    issueNumber,
-    useMultiAgentPipeline: useMultiAgent,
-  });
-  if (!useMultiAgent) {
-    logger.info('dispatchParallelImplement: pipeline disabled, skipping', { slug, issueNumber });
-    return;
-  }
-
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchParallelImplement: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchParallelImplement: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchParallelImplement);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runParallelImplementWorkflow } = (await import(sliceUrl('parallel-implement'))) as {
-      runParallelImplementWorkflow: (
-        item: unknown,
-        spec: unknown,
-        pipelineRunId: string,
-        source: unknown,
-        slug: string,
-        targetRepo: string,
-        deps?: Record<string, unknown>,
-      ) => Promise<{ status: 'success' | 'failed'; errorReason?: string }>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchParallelImplement: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:spec-ready') {
-      logger.info('dispatchParallelImplement: state already moved, skipping', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-
-    const specRecord = getEngineeringSpec(slug, item.id);
-    if (specRecord == null) {
-      await source.comment(
-        item.externalId,
-        'parallel-implement: no engineering spec found for this work item. Escalating to needs-human.',
-      );
-      await source.transitionState(item.externalId, 'factory:spec-ready', 'factory:needs-human');
-      emitStateTransitionEvent({
-        projectId: slug,
-        workItemId: item.id,
-        from: 'factory:spec-ready',
-        to: 'factory:needs-human',
-        by: 'parallel-implement',
-      });
-      return;
-    }
-
-    const parsedSpec = EngineeringSpecSchema.safeParse(specRecord.spec);
-    if (!parsedSpec.success) {
-      await source.comment(
-        item.externalId,
-        'parallel-implement: persisted engineering spec failed schema validation. Escalating to needs-human.',
-      );
-      await source.transitionState(item.externalId, 'factory:spec-ready', 'factory:needs-human');
-      emitStateTransitionEvent({
-        projectId: slug,
-        workItemId: item.id,
-        from: 'factory:spec-ready',
-        to: 'factory:needs-human',
-        by: 'parallel-implement',
-      });
-      return;
-    }
-
-    // Inject mock deps when running under the e2e harness so filesystem/git
-    // operations don't fail against non-existent worktrees.
-    const parallelMockDeps: Record<string, unknown> =
-      process.env.MOCK_AGENTS === 'true' || process.env.MOCK_OPEN_PR === 'true'
-        ? {
-            runtime: new ClaudeCliRuntime(),
-            openPRImpl: () =>
-              Promise.resolve({
-                prNumber: 999,
-                prUrl: 'https://github.com/shaunnez/goose-hub/pull/999',
-                branch: 'factory/mock-parallel-run',
-                base: 'main',
-              }),
-            createIssueWorktreeImpl: (_repo: string, runId: string) => {
-              const dir = join(tmpdir(), `e2e-parallel-issue-${runId}`);
-              mkdirSync(dir, { recursive: true });
-              return dir;
-            },
-            createWpWorktreeImpl: (_repo: string, runId: string, wpId: string) => {
-              const dir = join(tmpdir(), `e2e-parallel-wp-${runId}-${wpId}`);
-              mkdirSync(dir, { recursive: true });
-              return dir;
-            },
-            cleanupWpWorktreesImpl: () => {},
-            orchestratorCommitWpImpl: () => 'mock-commit-sha',
-            revertWpChangesImpl: () => {},
-            getDiffImpl: () => '',
-            commitDevReviewResponseImpl: () => 'mock-dr-commit-sha',
-          }
-        : {};
-
-    await source.transitionState(item.externalId, 'factory:spec-ready', 'factory:in-progress');
-    emitStateTransitionEvent({
-      projectId: slug,
-      workItemId: item.id,
-      from: 'factory:spec-ready',
-      to: 'factory:in-progress',
-      by: 'parallel-implement',
-      runId: specRecord.pipelineRunId,
-    });
-    try {
-      const result = await runParallelImplementWorkflow(
-        item,
-        parsedSpec.data,
-        specRecord.pipelineRunId,
-        source,
-        slug,
-        REPO_ROOT,
-        parallelMockDeps,
-      );
-      if (result.status === 'success') {
-        await source.transitionState(item.externalId, 'factory:in-progress', 'factory:needs-qa');
-        emitStateTransitionEvent({
-          projectId: slug,
-          workItemId: item.id,
-          from: 'factory:in-progress',
-          to: 'factory:needs-qa',
-          by: 'parallel-implement',
-          runId: specRecord.pipelineRunId,
-        });
-      } else {
-        await source.transitionState(item.externalId, 'factory:in-progress', 'factory:needs-human');
-        emitStateTransitionEvent({
-          projectId: slug,
-          workItemId: item.id,
-          from: 'factory:in-progress',
-          to: 'factory:needs-human',
-          by: 'parallel-implement',
-          runId: specRecord.pipelineRunId,
-        });
-      }
-    } catch (err) {
-      const error = err instanceof Error ? err : new Error(String(err));
-      await source.comment(
-        item.externalId,
-        `parallel-implement: workflow failed after dispatch. Escalating to needs-human.\n\nError: ${error.message}`,
-      );
-      await source.transitionState(item.externalId, 'factory:in-progress', 'factory:needs-human');
-      emitStateTransitionEvent({
-        projectId: slug,
-        workItemId: item.id,
-        from: 'factory:in-progress',
-        to: 'factory:needs-human',
-        by: 'parallel-implement',
-        runId: specRecord.pipelineRunId,
-      });
-    }
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/** Run the merge conflict resolution workflow for a single issue. Drops duplicate triggers. */
-export async function dispatchResolveConflict(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchResolveConflict: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchResolveConflict: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchResolveConflict);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runResolveConflictWorkflow } = (await import(sliceUrl('resolve-conflict'))) as {
-      runResolveConflictWorkflow: (
-        item: unknown,
-        source: unknown,
-        slug: string,
-        repoRoot: string,
-        deps?: Record<string, unknown>,
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchResolveConflict: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    const mockConflictDeps: Record<string, unknown> | undefined =
-      process.env.MOCK_SOURCE === 'true'
-        ? {
-            mergePRImpl: () => Promise.resolve({ sha: 'mock-sha-merged', merged: true }),
-            gitExecImpl: () => '',
-          }
-        : undefined;
-    await runResolveConflictWorkflow(item, source, slug, REPO_ROOT, mockConflictDeps);
-    // Fire-and-forget retro after conflict resolution + merge, same pattern as
-    // approveIssue. The label-change webhook also triggers dispatchRetro on
-    // factory:retrospecting; running it here avoids waiting for webhook delivery.
-    dispatchRetro(slug, issueNumber).catch((err: unknown) => {
-      logger.error('dispatchRetro after resolve-conflict failed', {
-        slug,
-        issueNumber,
-        error: String(err),
-      });
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/** Run the QA holdout workflow for a single issue. Drops duplicate triggers for the same issue. */
-export async function dispatchQa(slug: string, issueNumber: number): Promise<void> {
-  const projectForFlag = await getProject(slug);
-  const useMultiAgent =
-    projectForFlag != null ? getUseMultiAgentPipeline(projectForFlag.id) : false;
-  logger.info('dispatchQa: pipeline flag', {
-    slug,
-    issueNumber,
-    useMultiAgentPipeline: useMultiAgent,
-  });
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchQa: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchQa: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchQa);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runQaWorkflow } = (await import(sliceUrl('qa'))) as {
-      runQaWorkflow: (
-        item: unknown,
-        source: unknown,
-        projectSlug: string,
-        targetRepo: string,
-        deps?: {
-          verifyCommands?: Array<{
-            ac: string;
-            command: string;
-            expected: string;
-            tolerance: string;
-          }>;
-          runTests?: (() => Promise<null>) | undefined;
-        },
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchQa: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    const verifyCommands = parseAcceptanceCriteria(item.body ?? '');
-    const qaRunTests = process.env.MOCK_SOURCE === 'true' ? () => Promise.resolve(null) : undefined;
-    await runQaWorkflow(item, source, slug, item.repoRef ?? slug, {
-      verifyCommands,
-      runTests: qaRunTests,
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/** Run the fix-feedback workflow for a single issue. Drops duplicate triggers. */
-export async function dispatchNeedsFix(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchNeedsFix: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchNeedsFix: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchNeedsFix);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runFixFeedbackWorkflow } = (await import(sliceUrl('fix-feedback'))) as {
-      runFixFeedbackWorkflow: (
-        item: unknown,
-        source: unknown,
-        slug: string,
-        targetRepo: string,
-        deps?: Record<string, unknown>,
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchNeedsFix: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:needs-fix') {
-      logger.info('dispatchNeedsFix: state already moved, skipping', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-    await runFixFeedbackWorkflow(item, source, slug, item.repoRef ?? slug);
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Auto-transition from qa-failed based on retry logic.
- * QA already applied shouldEscalateQa — if we reach this label, retries
- * remain. Transition qa-failed → needs-fix and dispatch fix-feedback.
- */
-async function dispatchQaFailed(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchQaFailed: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchQaFailed: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchQaFailed);
-    return;
-  }
-  try {
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchQaFailed: no source for slug', { slug });
-      return;
-    }
-
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:qa-failed') {
-      logger.info('dispatchQaFailed: state already moved', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-
-    const workItemId = `github:${source.repoRef}#${issueNumber}`;
-    await source.transitionState(workItemId, 'factory:qa-failed', 'factory:needs-fix');
-
-    emitStateTransitionEvent({
-      projectId: slug,
-      workItemId,
-      from: 'factory:qa-failed',
-      to: 'factory:needs-fix',
-      by: 'orchestrator',
-    });
-
-    logger.info('dispatchQaFailed: transitioned to needs-fix, dispatching fix-feedback', {
-      slug,
-      issueNumber,
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-
-  // Fire fix-feedback outside the in-flight guard so needs-fix can also
-  // be triggered independently (e.g. review-failed path in the future).
-  await dispatchNeedsFix(slug, issueNumber);
-}
-
-/** Run the Review holdout workflow for a single issue. Drops duplicate triggers for the same issue. */
-export async function dispatchReview(slug: string, issueNumber: number): Promise<void> {
-  const projectForFlag = await getProject(slug);
-  const useMultiAgent =
-    projectForFlag != null ? getUseMultiAgentPipeline(projectForFlag.id) : false;
-  logger.info('dispatchReview: pipeline flag', {
-    slug,
-    issueNumber,
-    useMultiAgentPipeline: useMultiAgent,
-  });
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchReview: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchReview: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchReview);
-    return;
-  }
-  try {
-    // Cross-package boundary: slices/ is not a workspace package (rule 28a).
-    const { runReviewWorkflow, runConvergentReviewWorkflow } = (await import(
-      sliceUrl('review')
-    )) as {
-      runReviewWorkflow: (
-        item: unknown,
-        source: unknown,
-        projectSlug: string,
-        targetRepo: string,
-      ) => Promise<unknown>;
-      runConvergentReviewWorkflow: (
-        item: unknown,
-        source: unknown,
-        projectSlug: string,
-        targetRepo: string,
-      ) => Promise<unknown>;
-    };
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchReview: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-
-    const useConvergent = useMultiAgent;
-
-    if (useConvergent) {
-      await runConvergentReviewWorkflow(item, source, slug, item.repoRef ?? slug);
-    } else {
-      await runReviewWorkflow(item, source, slug, item.repoRef ?? slug);
-    }
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Run the retrospective workflow for a single issue. Drops concurrent
- * duplicates via the in-flight guard, AND skips items whose state already
- * advanced past factory:retrospecting — retro is triggered both directly
- * from approveIssue and via the factory:retrospecting label webhook, so
- * sequential duplicate triggers must no-op once the workflow has run.
- */
-export async function dispatchRetro(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchRetro: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchRetro: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchRetro);
-    return;
-  }
-  try {
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchRetro: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:retrospecting') {
-      logger.info('dispatchRetro: state already advanced, skipping', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-    await runRetroForItem(item, source, slug);
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Auto-transition from investigation-complete based on confidence.
- * Low confidence → gate-pending (human review required).
- * Medium/high → dev-ready (proceed automatically).
- */
-async function dispatchInvestigationComplete(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchInvestigationComplete: duplicate in-flight, dropping', {
-      slug,
-      issueNumber,
-    });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchInvestigationComplete: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchInvestigationComplete);
-    return;
-  }
-  try {
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchInvestigationComplete: no source for slug', { slug });
-      return;
-    }
-
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:investigation-complete') {
-      logger.info('dispatchInvestigationComplete: state already moved', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-
-    const workItemId = `github:${source.repoRef}#${issueNumber}`;
-    const allEvents = eventStore.replay({ projectId: slug, workItemId });
-    const investigationEvents = allEvents.filter((e) => e.kind === 'agent.investigation-complete');
-    const latest = investigationEvents.at(-1);
-    const confidence =
-      (latest?.payload as { investigate?: { confidence?: string } } | null)?.investigate
-        ?.confidence ?? 'medium';
-
-    const targetState = confidence === 'low' ? 'factory:gate-pending' : 'factory:dev-ready';
-    await source.transitionState(workItemId, 'factory:investigation-complete', targetState);
-
-    emitStateTransitionEvent({
-      projectId: slug,
-      workItemId,
-      from: 'factory:investigation-complete',
-      to: targetState,
-      by: 'orchestrator',
-    });
-
-    if (targetState === 'factory:gate-pending') {
-      eventStore.appendEvent({
-        projectId: slug,
-        workItemId,
-        kind: 'gate.awaiting-human',
-        payload: {
-          reason:
-            'Investigation confidence is low — human review required before proceeding to dev.',
-        },
-      });
-    }
-
-    logger.info('dispatchInvestigationComplete: transitioned', { slug, issueNumber, targetState });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Handles terminal labels (factory:archived, factory:rejected) that bypass the
- * retro path. Fetches the issue's milestone and fires maybeFireSprintReview so
- * that the sprint-review trigger runs even when an issue is archived or rejected
- * without going through retrospective (PLAN §12.6).
- */
-async function dispatchTerminalLabel(slug: string, issueNumber: number): Promise<void> {
-  const source = await getSourceForSlug(slug);
-  if (source == null) {
-    logger.error('dispatchTerminalLabel: no source for slug', { slug, issueNumber });
-    return;
-  }
-  const item = await source.getItem(issueNumber.toString());
-  const milestoneNumber = item.milestoneId != null ? Number(item.milestoneId) : null;
-  const milestoneTitle = item.milestoneTitle ?? null;
-  if (milestoneNumber == null || Number.isNaN(milestoneNumber) || milestoneTitle == null) {
-    logger.info('dispatchTerminalLabel: no milestoneId, skipping sprint-review check', {
-      slug,
-      issueNumber,
-    });
-    return;
-  }
-  // Fire-and-forget; errors logged inside maybeFireSprintReview.
-  void maybeFireSprintReview(slug, milestoneNumber, milestoneTitle, source);
-}
-
-/**
- * Webhook label-driven dispatcher. Routes the factory:* label to the right
- * workflow without requiring the webhook handler to know about the
- * `workflows` or `slices` directory layout.
- */
-export async function dispatchForLabel(
-  slug: string,
-  issueNumber: number,
-  labelName: string,
-): Promise<void> {
-  if (labelName === 'factory:triaging') {
-    await dispatchTriageBatch(slug);
-    return;
-  }
-  if (labelName === 'factory:investigating') {
-    await dispatchInvestigate(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:investigation-complete') {
-    await dispatchInvestigationComplete(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:dev-ready') {
-    await dispatchFixIssue(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:spec-ready') {
-    await dispatchParallelImplement(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:needs-qa') {
-    await dispatchQa(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:needs-review') {
-    await dispatchReview(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:merge-conflict') {
-    await dispatchResolveConflict(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:retrospecting') {
-    await dispatchRetro(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:qa-failed') {
-    await dispatchQaFailed(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:needs-fix') {
-    await dispatchNeedsFix(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:grilling') {
-    await dispatchGrillAndPrd(slug, issueNumber);
-    return;
-  }
-  if (labelName === 'factory:decomposing') {
-    await dispatchDecomposePrd(slug, issueNumber);
-    return;
-  }
-  // Terminal labels that bypass retro — check for sprint-review eligibility.
-  if (labelName === 'factory:archived' || labelName === 'factory:rejected') {
-    await dispatchTerminalLabel(slug, issueNumber);
-    return;
-  }
-  logger.info('dispatchForLabel: no workflow for label', { slug, labelName });
-}
-
-// ─── M13 Discover Lane dispatchers ────────────────────────────────────────────
-
-const PRD_MARKER = '<!-- factory:prd -->';
-const GRILL_QUESTION_MARKER = '<!-- factory:grill-question -->';
-const SYSTEM_MARKER = '<!-- factory:system -->';
-const CHILD_ISSUES_MARKER = '## Child issues';
-const PRD_JSON_FENCE_RE = /```json\s*\n([\s\S]*?)\n```/;
-
-/**
- * Extract the PRDOutput JSON from the latest `<!-- factory:prd -->` marker
- * comment. Returns `null` if no marker comment exists or the JSON fence
- * fails to parse. The decompose-prd workflow validates the parsed shape
- * against `DecomposeOutputSchema`'s upstream `PRDOutputSchema` itself.
- */
-function extractPrdFromComments(
-  comments: ReadonlyArray<{ body: string; createdAt: string }>,
-): unknown {
-  const matches = comments.filter((c) => c.body.startsWith(PRD_MARKER));
-  if (matches.length === 0) return null;
-  const sorted = [...matches].sort((a, b) => (a.createdAt < b.createdAt ? -1 : 1));
-  const fence = sorted[sorted.length - 1].body.match(PRD_JSON_FENCE_RE);
-  if (fence == null) return null;
-  try {
-    return JSON.parse(fence[1]);
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Build the `priorReplies` array for grill-and-prd from issue comments.
- * Agent messages: grill questions (<!-- factory:grill-question -->) and PRD
- * drafts (<!-- factory:prd -->). PRD drafts are included so the griller sees
- * what was produced when a user declines and re-enters grilling.
- * System-marker and child-issues comments are excluded as noise.
- */
-function buildPriorReplies(
-  comments: ReadonlyArray<{ body: string }>,
-): Array<{ role: 'user' | 'agent'; content: string }> {
-  return comments
-    .filter((c) => !c.body.startsWith(SYSTEM_MARKER) && !c.body.startsWith(CHILD_ISSUES_MARKER))
-    .map((c) => ({
-      role:
-        c.body.startsWith(GRILL_QUESTION_MARKER) || c.body.startsWith(PRD_MARKER)
-          ? ('agent' as const)
-          : ('user' as const),
-      content: c.body,
-    }));
-}
-
-/**
- * Run the grill-and-prd workflow for a single issue. Drops duplicate
- * triggers via parallel-lock. Webhook fires this when the issue lands in
- * `factory:grilling`; the rejectPRD handler fires it after returning a
- * rejected PRD to grilling.
- */
-export async function dispatchGrillAndPrd(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchGrillAndPrd: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchGrillAndPrd: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchGrillAndPrd);
-    return;
-  }
-  try {
-    const { runGrillAndPrdWorkflow } = await import('@goose-hub/core/workflows/grill-and-prd.js');
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchGrillAndPrd: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:grilling') {
-      logger.info('dispatchGrillAndPrd: state already advanced, skipping', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-    const comments = await source.listComments(issueNumber.toString());
-    const priorReplies = buildPriorReplies(comments);
-    const mockGrillDeps =
-      process.env.MOCK_AGENTS === 'true'
-        ? {
-            createWorktreeImpl: (_repo: string, _runId: string) => '/mock/worktree',
-            cleanupWorktreeImpl: (_runId: string) => undefined,
-          }
-        : undefined;
-    await runGrillAndPrdWorkflow({
-      workItem: item,
-      stateSource: source,
-      projectId: slug,
-      priorReplies,
-      deps: { repoRoot: REPO_ROOT, ...(mockGrillDeps ?? {}) },
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Retry write-prd directly, skipping re-grilling. Used when an issue is stuck
- * in factory:prd-drafting after a failed or incomplete write-prd run. Grill
- * context (refinedIntent) is recovered from the grill.completed event in the
- * event store; priorReplies are rebuilt from issue comments so the PRD writer
- * has the full Q&A history.
- */
-export async function dispatchRetryWritePrd(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchRetryWritePrd: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchRetryWritePrd: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchRetryWritePrd);
-    return;
-  }
-  try {
-    // Case 2: runtime-resolved path — module path depends on per-project config loaded at runtime.
-    const { runGrillAndPrdWorkflow } = await import('@goose-hub/core/workflows/grill-and-prd.js');
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchRetryWritePrd: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:prd-drafting') {
-      logger.info('dispatchRetryWritePrd: state already advanced, skipping', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-    const comments = await source.listComments(issueNumber.toString());
-    const priorReplies = buildPriorReplies(comments);
-    await runGrillAndPrdWorkflow({
-      workItem: item,
-      stateSource: source,
-      projectId: slug,
-      priorReplies,
-      skipGrill: true,
-      deps: { repoRoot: REPO_ROOT },
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Re-run write-prd with prior PRD + human concerns (revise path).
- * Drops duplicate triggers via parallel-lock. State stays prd-review.
- */
-export async function dispatchRevisePrd(
-  slug: string,
-  issueNumber: number,
-  priorPrd: import('@goose-hub/skills/write-prd/schema.js').PRDOutput | undefined,
-  humanConcerns: string[],
-): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.warn('dispatchRevisePrd: parallel-lock rejected (in-flight or at cap)', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    return;
-  }
-  try {
-    const { runGrillAndPrdWorkflow } = await import('@goose-hub/core/workflows/grill-and-prd.js');
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchRevisePrd: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    const comments = await source.listComments(issueNumber.toString());
-    const priorReplies = buildPriorReplies(comments);
-    await runGrillAndPrdWorkflow({
-      workItem: item,
-      stateSource: source,
-      projectId: slug,
-      priorReplies,
-      priorPrd,
-      humanConcerns,
-      deps: { repoRoot: REPO_ROOT },
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-  }
-}
-
-/**
- * Run the decompose-prd workflow for a single issue. Drops duplicate
- * triggers via parallel-lock. Webhook fires this when the issue lands
- * in `factory:decomposing`; the approvePRD handler fires it after the
- * human approves the PRD comment.
- */
-export async function dispatchDecomposePrd(slug: string, issueNumber: number): Promise<void> {
-  const maxParallel = await getMaxParallelAgents(slug);
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchDecomposePrd: duplicate in-flight, dropping', { slug, issueNumber });
-    return;
-  }
-  if (!parallelLock.tryAcquire(slug, issueNumber, maxParallel)) {
-    logger.info('dispatchDecomposePrd: cap full, queuing work item', {
-      slug,
-      issueNumber,
-      inFlight: parallelLock.inFlightCount(slug),
-      maxParallel,
-    });
-    enqueueWorkflow(slug, issueNumber, dispatchDecomposePrd);
-    return;
-  }
-  try {
-    const { runDecomposePrdWorkflow } = await import('@goose-hub/core/workflows/decompose-prd.js');
-    const source = await getSourceForSlug(slug);
-    if (source == null) {
-      logger.error('dispatchDecomposePrd: no source for slug', { slug });
-      return;
-    }
-    const item = await source.getItem(issueNumber.toString());
-    if (item.state !== 'factory:decomposing') {
-      logger.info('dispatchDecomposePrd: state already advanced, skipping', {
-        slug,
-        issueNumber,
-        state: item.state,
-      });
-      return;
-    }
-    const comments = await source.listComments(issueNumber.toString());
-    const prdOutput = extractPrdFromComments(comments);
-    if (prdOutput == null) {
-      logger.error('dispatchDecomposePrd: no PRD marker comment found', { slug, issueNumber });
-      await source.comment(
-        issueNumber.toString(),
-        'decompose-prd: no PRD comment found on this issue. Returning to needs-human.',
-      );
-      await source.forceState(issueNumber.toString(), 'factory:needs-human');
-      return;
-    }
-    await runDecomposePrdWorkflow({
-      workItem: item,
-      prdOutput,
-      stateSource: source,
-      projectId: slug,
-    });
-  } finally {
-    parallelLock.release(slug, issueNumber);
-    drainPending(slug);
-  }
-}
-
-/**
- * Dispatch the workflow that matches an issue's current factory:* state.
- * Webhook-free escape hatch used by e2e tests so they don't depend on
- * GitHub webhook delivery to the local server.
- */
-export async function dispatchForIssue(slug: string, issueNumber: number): Promise<void> {
-  const source = await getSourceForSlug(slug);
-  if (source == null) {
-    logger.error('dispatchForIssue: no source for slug', { slug });
-    return;
-  }
-  const item = await source.getItem(issueNumber.toString());
-  await dispatchForLabel(slug, issueNumber, item.state);
-}
-
-type ResumeEntry = {
-  targetState: StateName;
-  dispatch: (slug: string, issueNumber: number) => Promise<void>;
-};
-
-const RESUME_WORKFLOWS: Partial<Record<StateName, ResumeEntry>> = {
-  // Re-run the full triage batch so a stalled triage or repo-match run is retried.
-  'factory:triaging': {
-    targetState: 'factory:triaging',
-    dispatch: (slug: string, _issueNumber: number) => dispatchTriageBatch(slug),
-  },
-  'factory:dev-ready': { targetState: 'factory:dev-ready', dispatch: dispatchFixIssue },
-  'factory:spec-ready': { targetState: 'factory:spec-ready', dispatch: dispatchParallelImplement },
-  'factory:in-progress': { targetState: 'factory:dev-ready', dispatch: dispatchFixIssue },
-  'factory:needs-qa': { targetState: 'factory:needs-qa', dispatch: dispatchQa },
-  'factory:needs-review': { targetState: 'factory:needs-review', dispatch: dispatchReview },
-  'factory:merge-conflict': {
-    targetState: 'factory:merge-conflict',
-    dispatch: dispatchResolveConflict,
-  },
-  'factory:investigating': { targetState: 'factory:investigating', dispatch: dispatchInvestigate },
-  'factory:retrospecting': { targetState: 'factory:retrospecting', dispatch: dispatchRetro },
-  'factory:qa-failed': { targetState: 'factory:needs-fix', dispatch: dispatchNeedsFix },
-  'factory:needs-fix': { targetState: 'factory:needs-fix', dispatch: dispatchNeedsFix },
-  // Discover lane
-  'factory:grilling': { targetState: 'factory:grilling', dispatch: dispatchGrillAndPrd },
-  // factory:gate-pending is handled above with lane-origin inspection; it is
-  // intentionally absent from this table so the special-case runs first.
-  'factory:prd-drafting': { targetState: 'factory:prd-drafting', dispatch: dispatchRetryWritePrd },
-  'factory:decomposing': { targetState: 'factory:decomposing', dispatch: dispatchDecomposePrd },
-};
-
-/**
- * Resume an orphaned or stalled run. Looks up the issue's current state,
- * forces it back to the canonical trigger state for that workflow (e.g.
- * factory:in-progress → factory:dev-ready), then re-dispatches. Uses
- * forceState rather than transitionState because recovery transitions are
- * not always legal in the normal workflow graph.
- *
- * factory:gate-pending is handled specially: the last `state.transitioned`
- * event that landed the issue in gate-pending is inspected for its `from`
- * field to determine which lane to resume. Only `from: factory:grilling`
- * supports auto-resume (routes back to the grill workflow). All other origins
- * require manual triage.
- */
-export async function dispatchResumeIssue(slug: string, issueNumber: number): Promise<void> {
-  if (parallelLock.isInFlight(slug, issueNumber)) {
-    logger.warn('dispatchResumeIssue: already in-flight, dropping duplicate', {
-      slug,
-      issueNumber,
-    });
-    return;
-  }
-  const source = await getSourceForSlug(slug);
-  if (source == null) {
-    logger.error('dispatchResumeIssue: no source for slug', { slug });
-    return;
-  }
-  const workItemId = `github:${source.repoRef}#${issueNumber}`;
-  const item = await source.getItem(issueNumber.toString());
-  const fromState = item.state;
-
-  // gate-pending is lane-agnostic. Inspect the last state.transitioned event
-  // that moved the issue into gate-pending to determine which lane to resume.
-  if (fromState === 'factory:gate-pending') {
-    const allEvents = eventStore.replay({ projectId: slug, workItemId });
-    const lastToGatePending = [...allEvents]
-      .reverse()
-      .find(
-        (e) =>
-          e.kind === 'state.transitioned' &&
-          (e.payload as { to?: string }).to === 'factory:gate-pending',
-      );
-    const transitionedFrom = (lastToGatePending?.payload as { from?: string } | undefined)?.from;
-
-    if (transitionedFrom === 'factory:grilling') {
-      logger.info('dispatchResumeIssue: gate-pending from grilling, resuming discover lane', {
-        slug,
-        issueNumber,
-      });
-      await source.forceState(workItemId, 'factory:grilling');
-      emitStateTransitionEvent({
-        projectId: slug,
-        workItemId,
-        from: fromState,
-        to: 'factory:grilling',
-        by: 'resume',
-      });
-      await dispatchGrillAndPrd(slug, issueNumber);
-      return;
-    }
-
-    logger.warn('dispatchResumeIssue: gate-pending from unknown lane, cannot auto-resume', {
-      slug,
-      issueNumber,
-      transitionedFrom,
-    });
-    return;
-  }
-
-  const entry = RESUME_WORKFLOWS[fromState];
-  if (entry == null) {
-    logger.warn('dispatchResumeIssue: no resume handler for state', {
-      slug,
-      issueNumber,
-      fromState,
-    });
-    return;
-  }
-
-  // If the issue reached in-progress via spec-ready it was launched by the parallel-implement
-  // pipeline. Resume by restoring spec-ready and re-dispatching (clean restart — no WP-level
-  // checkpoint recovery across server restarts). If it arrived via dev-ready, fall through to
-  // the legacy fix-issue path in RESUME_WORKFLOWS.
-  if (fromState === 'factory:in-progress') {
-    const allEvents = eventStore.replay({ projectId: slug, workItemId });
-    const lastToInProgress = [...allEvents]
-      .reverse()
-      .find(
-        (e) =>
-          e.kind === 'state.transitioned' &&
-          (e.payload as { to?: string }).to === 'factory:in-progress',
-      );
-    const arrivedFrom = (lastToInProgress?.payload as { from?: string } | undefined)?.from;
-
-    if (arrivedFrom === 'factory:spec-ready') {
-      logger.info('dispatchResumeIssue: in-progress via M19, resuming from spec-ready', {
-        slug,
-        issueNumber,
-      });
-      await source.forceState(workItemId, 'factory:spec-ready');
-      emitStateTransitionEvent({
-        projectId: slug,
-        workItemId,
-        from: fromState,
-        to: 'factory:spec-ready',
-        by: 'resume',
-      });
-      await dispatchParallelImplement(slug, issueNumber);
-      return;
-    }
-    // Legacy path: fall through to RESUME_WORKFLOWS entry (dev-ready → fix-issue)
-  }
-
-  if (entry.targetState !== fromState) {
-    await source.forceState(workItemId, entry.targetState);
-    emitStateTransitionEvent({
-      projectId: slug,
-      workItemId,
-      from: fromState,
-      to: entry.targetState,
-      by: 'resume',
-    });
-    logger.info('dispatchResumeIssue: forced state for resume', {
-      slug,
-      issueNumber,
-      fromState,
-      targetState: entry.targetState,
-    });
-  }
-
-  logger.info('dispatchResumeIssue: dispatching workflow', {
-    slug,
-    issueNumber,
-    state: entry.targetState,
-  });
-  await entry.dispatch(slug, issueNumber);
-}
+export { dispatchTriageBatch } from './dispatch-triage.js';
+export {
+  dispatchInvestigate,
+  dispatchSpecAuthor,
+  dispatchFixIssue,
+  dispatchParallelImplement,
+  dispatchResolveConflict,
+} from './dispatch-dev.js';
+export {
+  dispatchQa,
+  dispatchReview,
+  dispatchRetro,
+  dispatchNeedsFix,
+} from './dispatch-qa-review.js';
+export {
+  dispatchGrillAndPrd,
+  dispatchRetryWritePrd,
+  dispatchRevisePrd,
+  dispatchDecomposePrd,
+} from './dispatch-discover.js';
+export { dispatchForLabel, dispatchForIssue, dispatchResumeIssue } from './dispatch-routing.js';
